+++ src/home/routes.ts
@@ -1,11 +1,5 @@
 import { Router, Request, Response } from "express";
 import PromiseRouter from "express-promise-router";
-import { IContactRequest } from "../contact/contactRequest";
-import { recaptcha, recaptchaSiteKey } from "../shared/recaptcha";
-import { RecaptchaResponseDataV3 } from "express-recaptcha/dist/interfaces";
-import formData from "form-data";
-import Mailgun from "mailgun.js";
-import config from "config";
 import { Music } from "../music/musicService";
 import { formatDate, formatPieceYear, formatYear } from "../shared/dateHelpers";
 
@@ -30,45 +24,8 @@ router.get("/cart", (_req: Request, res: Response) => {
   res.render("cart", { title: "Cart" });
 });
 
-router.get("/contact", (_req: Request, res: Response) => {
-  res.render("contact", { title: "Contact", recaptchaSiteKey });
-});
-
 router.get("/privacy-policy", (_req: Request, res: Response) => {
   res.render("privacy-policy", { title: "Privacy Policy"});
-});
-
-router.post("/api/contact", recaptcha.middleware.verify, async (req: Request, res: Response) => {
-  const data: IContactRequest = req.body;
-
-  if (!req.recaptcha ||
-      req.recaptcha.error ||
-      !((req.recaptcha.data as RecaptchaResponseDataV3).score) ||
-      (req.recaptcha.data as RecaptchaResponseDataV3).score < 0.5) {
-    res.sendStatus(400);
-    return;
-  }
-
-  const mailgun = new Mailgun(formData);
-  const mailgunClient = mailgun.client({ key: config.get<string>("mailgunApiKey"), username: "api" });
-
-  const message = {
-    to: [config.get<string>("contactToEmail")],
-    from: "Contact Form <contact@alexander-roode.com>",
-    subject: "Contact Form Submission",
-    html: `<strong>Name</strong>: ${data.name}<br/>` +
-    `<strong>Email</strong>: ${data.email}<br/>` +
-    `<strong>Message</strong>: <br/><p>${data.message}</p>`
-  };
-
-  try {
-    await mailgunClient.messages
-      .create(config.get<string>("mailgunDomain"), message);
-    return res.sendStatus(204);
-  } catch (error) {
-    console.error(error);
-    res.sendStatus(500);
-  }
 });
 
 export const HomeRoutes: Router = router;
